requirements.txt
# Core ML & Data Processing
pandas>=2.1.0
numpy>=1.24.0

# AI Agent Framework (LangChain)
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-experimental>=0.3.0

# Data Validation & Environment
pydantic>=2.0.0
python-dotenv>=1.0.1

# Utilities
tabulate>=0.9.0